Solve for async initial State for application
Complicated and need to refactor.  Also need crud actions to go to
database vs local storage at app close for persistence

Needed to connect stor to the TodoList component and allow for
componentDidMount to facilitate an async call for data.  The call
then pulls from db and reducer resets the initial state to the new
data pulled.

Also need to display error and loading messages in the app