Final fix for datePicker saving state
Was not using correct variable name in mapDispatch so action creator
was not being called.  Truly need to set up front end testing to
solve for this stuff as well.  Dates working properly in the addTodo
and the Update Todo areas.  TodoItem is a connected component, but
still need to finish the other CRUD operations migration.