Add todo form not saving refs and todo.focus not a function
Used git diff commit1 and commit2 to quickly debug the files and see
the impacting changes.  commit1 was the earlier version and the `+`
signs indicated additions, etc.  Had started to checkout different
branches with git checkout <branch> to see which was the last
working version and then applies the git diff.  Saved a lot of time.