Homepage loads initial tasks from api
Needed to follow a formula for changing state
including:
- set up an apiAction creator
- build a reducer for the api actions
- make the homepage a store connection
- import and call the action fn from mapTo
- utilizing fetch vs axios (tutorial boilerplate)

Initial store populates from the api (on heroku)
which in turn calls to mlabs online database.

The raw data there is polluted with extraneous test
suite call objects and needs to be cleaned

Consider removing the APP landing page and returning
the title from HomePage component instead (no need)
for a landing page as loading is fast).