Finish Edit control for TodoTask
- refactored into a stateful component for TodoTask
- logic for completed status
- selects all text and changes styles for editing
- changes editing state accordingly
- blocks editing via alert for completed task
- toggles completed style based on parent default styling logic
- updates the item state based on changes
- handle methods bound to object via => function
- test and warnings are cleared

Stuggled for a while with basic flow of props for functional and
stateless components and how to get UI to reflect data.
Some Key learnings:

- passing props to a manufacture component- still need to set props
values in the file of the child component. e.g. if passing 'style'
from the parent as a prop, in the child component file, need to set
'style' prop to `this.props.style`.  Otherwise, no data will flow

- important to use component lifecyle to ensure state and props
update for stateful components not directly connected to store via
subscription.  componentWillMount for initial status and largely
componentWillReceiveProps to ensure props update accordingly. Other
wise run risk of disjointed state changes to UI.

- Be careful of the `this` statement in stateful components and
ensure that variables and functions are placed logically relative
to state declarations.

- When in doubt, take a break and think of the fundamental flow of
state and data. A few minutes for a quick review can save time.

-