UI Create operations posting to db and to the app state
Struggled with this one.  Appears the axios packaged updated its API
and broke the flow of data back from the server, therefor the
create actions addTodo fn was not getting data passed to it so that
it could update the UI.  Postman showed server was not the issue, so
could only see updated items after a refresh of the page.

res.data was not be captured in proper part of the Promise flow and
an if/else wrapper also appeared to be important in the createTodo fn