Initial todos loading on header click with Bearer Authentication
Needed to fix promise chain in the axios get path.  Can still
create todos and they now load based on authenticated user (filtered)

CRUD actions in UI appear to be functioning normally.  Setting to
axios allows for the default header options to be applied to all of
the axios routes, thus authenticating.