Add styled-components and update Home.js
Added.  In order for custom comonents to  map to grid areas,
need to ensure className is passed as a prop to all children.

Stamps won't automatically fill in the named area 'stamps'.
Need to assign each to a named area or set the start/end rows.

Learning curve here but useful for component architecture.