Complete coding of mobile drop down menu
designed a visible version of the grid display for mobile and then
employed ::hover to hide and restyle the elements on page vs
creation of a separate new unit.

For ::hover to allow access to a men item, menu must be a child of
the triggeringelement so that it stays in position until a selection
is made.

The `~` selctor was helpful for siblings, but went with the child
strategy outlined above.  Now to fix full size view and small
refactor, then publish.